Rename copy file to be language agnostic
To provide i18n support, gettext will be used so language specific
modules are not needed; rename the language file to be neutral.
