+++ kano_updater_gui/header.py
@@ -11,7 +11,7 @@
 from gi.repository import Gtk
 
 from kano_updater_gui.paths import IMAGE_PATH
-from kano_updater_gui.text_en import STATUS_TITLES, STAGE_TITLES, STAGE_BODIES
+from kano_updater_gui.text import STATUS_TITLES, STAGE_TITLES, STAGE_BODIES
 
 
 class Header(Gtk.EventBox):
