+++ kano_updater_gui/text.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-# text_en.py
+# text.py
 #
 # Copyright (C) 2015 Kano Computing Ltd.
 # License: http://www.gnu.org/licenses/gpl-2.0.txt GNU General Public License v2
